learn/requirements.txt
matplotlib==3.1.1
Pillow==6.2.1
imutils==0.5.3
keras==2.3.1
np==1.0.2
opencv-python==4.1.1.26
prometheus-async==19.2.0
scikit-image==0.16.2
scikit-learn==0.21.3
tensorflow==1.14.0
testresources==2.0.1